epics/requirements.txt
python-dateutil
PyYaml
requests
scipy
tqdm
